Moved logic for a copy of the operand buffers into the Vault Controller.
To do this, I had to add a copy of the operandBufID field held in each
packet for the Vault Controller's copy of the operand buffer (while testing
to see that it will work). This is the VoperandBufID field. It is
remembered when the VoperandBufID is chosen as the ACT_ADD comes into the
Vault Controller and used when the pcuPacket is ready to be sent up
and removed, when the operandBuffer entry is freed.