Flow Table Partially Replicated
Flow table is almost copied into the vault controller... The GET request is sent
to only one of the vaults (marked when the flow is created in the cube where there
is a vault controller being used). Both the operand entries and flow table entries
are created in the VC, but only the operand entries are consumed for now.