Added support for (simple) MULT trees at the vault level.
Now, Crossbar Switch has access to
the vault controllers to query them for how many operand entries are left.
If there is an operand buffer, it is reserved for this MULT operation before
the Crossbar Switch makes the requests to the other vaults. If those requests
go to other cubes, it is not supported so far, so there will likely be an
assertion failure. Note that now, VaultController.cpp has to be compiled
before CrossbarSwitch.cpp.