Refactored a way to collect network stats
Now, there's a (inefficient) way to collect stats from every vault in the
network based on the counts during currentClockCycle. Assuming each call to
Update() in an HMC happens on one clock cycle (according to currentClockCycle):

Each vault collects its stats for that Update() call, and the HMC sums them up
when it finishes calling Update() on the vaults. Then, it stores the total for
that cycle in a hash of cycle -> count.

When the network is about to destroy the HMCs, it goes through every cycle to
check if that hash has a key with that cycle number in every HMC. It sums up
all the counts of the HMCs that do have that key and increments the
histograms for that frequency in a cycle.